Fixed #66. If scan hasn't started, show '---' instead of 44 years ago or never.
Never as a default value sounds weird, so I proposed the switch to '---'.
Since the original implementation append a time elapse filter with ()
in the template, to make use of the changes in the filter
``scan_datetime`` without hardcoding the ``()`` in the Javascript (that's a
presentation logic), I use ``ng-show``.