Fixed #93. Send login once invite is accepted to get last_login date.
This fix might seem a waste for making a round trip. Well, the backend's
update_invite is monster and should be refactored to individual functions
to update both invite and user document. That being said, it isn't that
bad to make an extra round trip. It is a quick fix and actually does
reuse what has been implemented (login function in the frontend views
module).