Cleanup templates by centralizing nav context creation.
Initially looked into the low hanging fruit of Issue #86, but
decided it makes sense to do some refactoring here, to make the templates more
DRY, and provide sanity moving forward. Primary adjustments are:

* Compile all routes into an app.navItems variable
* Introduce app.navContext, allowing us to easily get the navigation
    items for a specific "section" of the site
* Adjust the controller to contain more variables that are useful
    in rendering a nav.

I am completely new to AngularJS, infact -- this is the first time I ever
touched it! So, there are some issues. For example, in `AdminGroupController`
we need a sub nav for the "Group Editor". The implementation of adding
this to the nav feels very hackish and perhaps we can implement
some notion of sub navigation. However, this may not be required right now
given the simplicity of the admin, so perhaps it's ok for the current
implementation.